+++ src/main/java/com/example/wizytydomowe/Patient/PatientRepository.java
@@ -3,5 +3,5 @@ package com.example.wizytydomowe.Patient;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PatientRepository extends JpaRepository<Patient, Integer> {
+public interface PatientRepository extends JpaRepository<Patient, Long> {
 }
